pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcs-release-monitor"
version = "0.1.0"
description = "Poll a Google Cloud Storage bucket, mirror new artifacts to Nextcloud, and emit signed webhook release events"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-cloud-storage>=2.18.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
gcs-release-monitor = "gcs_release_monitor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
